requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
gTTS==2.2.3
idna==3.2
pdf2image==1.16.0
Pillow==8.3.1
platformdirs==2.1.0
playsound==1.3.0
pydub==0.25.1
pytesseract==0.3.8
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
virtualenv==20.6.0
fastapi==0.68.0
uvicorn==0.14.0
python-multipart==0.0.5
aiofiles==0.7.0
apscheduler==3.7.0